requirements.txt
alembic==1.4.3
aniso8601==7.0.0
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
promise==2.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.19
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
